+++ voting/frontend/app/admin/page.tsx
@@ -1,19 +1,18 @@
 "use client";
 
-import { useAccount } from "wagmi";
-import { prepareWriteContract, readContract, writeContract } from "@wagmi/core";
-import { useEffect, useState } from "react";
-import { abi, contractAddress } from "@/constants";
-import { BaseError, ContractFunctionRevertedError } from "viem";
+import { useAccount } from "wagmi"
+import { useEffect, useState } from "react"
+import { readContractByFunctionName } from "@/utils"
+
+import VoterManager from "@/components/admin/VoterManager"
+import IsConnected from "@/components/IsConnected";
+import Event from "@/components/Event";
 
 const Admin = () => {
   const { address, isConnected } = useAccount()
   const [owner, setOwner] = useState(null)
   const [winningProposalID, setWinningProposalID] = useState(null)
   const [workflowStatus, setWorkflowStatus] = useState(0)
-  const [voter, setVoter] = useState('')
-  const [error, setError] = useState('');
-  const [sucess, setSucess] = useState('');
 
   const WorkflowStatus: string[] = [
     "RegisteringVoters",
@@ -23,14 +22,6 @@ const Admin = () => {
     "VotingSessionEnded",
     "VotesTallied"
   ]
-
-  const readContractByFunctionName = async (functionName: string): Promise<any> => {
-    return await readContract({
-      address: contractAddress,
-      abi: abi,
-      functionName: functionName,
-    });
-  }
 
   const getOwner = async () => {
     try {
@@ -59,42 +50,6 @@ const Admin = () => {
     }
   }
 
-  const addVoter = async () => {
-    const validAddress = new RegExp("^0x[a-fA-F0-9]{40}$")
-    setError('')
-    setSucess('')
-
-    if (!validAddress.test(voter)) {
-      setError('Invalid etherum address');
-    } else {
-      try {
-        const { request } = await prepareWriteContract({
-          address: contractAddress,
-          abi: abi,
-          functionName: 'addVoter',
-          args: [voter]
-        })
-
-        const { hash } = await writeContract(request)
-        setSucess(hash)
-      } catch (err: any) {
-        if (err instanceof BaseError) {
-          // Option 1: checking the instance of the error
-          if (err.cause instanceof ContractFunctionRevertedError) {
-            const cause: ContractFunctionRevertedError = err.cause
-            setError(cause.data?.errorName ?? 'Unkown error');
-          }
-
-          // Option 2: using `walk` method from `BaseError`
-          const revertError: any = err.walk(err => err instanceof ContractFunctionRevertedError)
-          if (revertError) {
-            setError(revertError.data?.message ?? 'Unkown error');
-          }
-        }
-      }
-    }
-  }
-
   useEffect(() => {
     if (isConnected) {
       getOwner()
@@ -108,9 +63,9 @@ const Admin = () => {
   }
 
   return (
-    <div className="flex flex-col space-y-2 mt-5 mx-auto max-w-screen-lg">
-      {isConnected ? (
-        isOwner() ? (
+    <div className="flex flex-col space-y-2 mx-auto max-w-screen-lg">
+      <IsConnected>
+        {isOwner() ? (
           <>
             <div className="mx-auto w-3/4 rounded h-auto bg-gradient-to-r from-indigo-900 to-indigo-600 text-indigo-100 shadow-lg">
               <div className="p-4">
@@ -132,29 +87,10 @@ const Admin = () => {
             </div>
 
             {WorkflowStatus[workflowStatus] === "RegisteringVoters" ? (
-              <div className="mx-auto w-3/4 rounded h-auto bg-gradient-to-r from-lime-300 to-lime-200 text-blue-gray-700 shadow-lg">
-                <div className="flex justify-between p-3">
-                  <div className="m-auto w-1/4">
-                    Add voter
-                  </div>
-
-                  <div className="m-auto w-2/4">
-                    <input type="text" onChange={e => setVoter(e.target.value)} className="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-yellow-700 focus:border-yellow-700 block w-full p-2" placeholder="Voter address" />
-                  </div>
-
-                  <div className="m-auto w-1/4 text-center">
-                    <button onClick={() => addVoter()} className="content-center bg-yellow-700 hover:bg-yellow-900 text-white font-semibold py-2 px-4 rounded-full">
-                      Add Voter
-                    </button>
-                  </div>
-                </div>
-                {error && <div className="text-red-800 font-semibold p-4 text-center">{error}</div>}
-                {sucess && <div className="text-green-600 font-semibold p-4 text-center">
-                  Voter added with sucess ! Transaction: {sucess}
-                </div>}
-              </div>
+              <VoterManager />
             ) : (<></>)}
 
+            <Event name='VoterRegistered'></Event>
           </>
 
         ) : (
@@ -175,14 +111,8 @@ const Admin = () => {
               </div>
             </div>
           </div>
-        )
-      ) : (
-        <div className="mx-auto w-3/4 rounded h-auto min-h-[50px] text-center bg-gradient-to-r from-indigo-950 to-rose-600 text-zinc-200 shadow-lg">
-          <div className="p-6 font-semibold text-md">
-            Please connect your Wallet.
-          </div>
-        </div>
-      )}
+        )}
+      </IsConnected>
     </div>
   )
 }