+++ voting/frontend/interfaces/Voter.ts
@@ -1,0 +1,5 @@
+export interface Voter {
+    hasVoted: boolean
+    isRegistered: boolean
+    votedProposalId?: number
+}